add interface builder support for variant-master
