+++ packages/app-schema-editor/interface-code-builder/index.ts
@@ -1,0 +1,1 @@
+export * from "./props-builder";
