+++ packages/app-schema-editor/interface-code-builder/props-builder.ts
@@ -1,0 +1,65 @@
+import { variant } from "@design-sdk/figma/features";
+import {
+  FigmaBoolean,
+  FigmaNumber,
+  FigmaUnique,
+  _FigmaVariantPropertyCompatType_to_string,
+} from "@design-sdk/figma/features/variant";
+import {
+  InterfaceDeclaration,
+  PropertySignature,
+  Identifier,
+  LiteralType,
+  StringLiteral,
+  UnionType,
+  BooleanKeyword,
+  NumberKeyword,
+} from "coli";
+
+import { stringfy } from "@coli.codes/export-string";
+
+export interface InterfaceCodeBuilParam {
+  name: string;
+  properties: {
+    name: string;
+    type: variant.FigmaVariantPropertyCompatType;
+  }[];
+}
+
+export function buildInterface(p: InterfaceCodeBuilParam) {
+  const _make_type = (t: variant.FigmaVariantPropertyCompatType) => {
+    if (t == FigmaNumber) {
+      return new NumberKeyword();
+    } else if (t == FigmaBoolean) {
+      return new BooleanKeyword();
+    } else if (t.type == "unique") {
+      return new LiteralType(new StringLiteral(t.value));
+    } else if (t.type == "enum") {
+      return new UnionType({
+        types: t.values.map((v) => {
+          return new LiteralType(new StringLiteral(v));
+        }),
+      });
+    }
+  };
+
+  return new InterfaceDeclaration({
+    name: p.name,
+    members: p.properties.map((n) => {
+      return new PropertySignature({
+        name: new Identifier(n.name),
+        type: _make_type(n.type),
+      });
+    }),
+  });
+}
+
+export function buildInterfaceString(p: InterfaceCodeBuilParam): string {
+  return interfaceColiToString(buildInterface(p));
+}
+
+export function interfaceColiToString(i: InterfaceDeclaration): string {
+  return stringfy(i, {
+    language: "typescript",
+  });
+}
